multi-modal-rag/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.1
ffmpeg-python==0.2.0
filelock==3.15.3
fonttools==4.53.0
fsspec==2024.6.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.23.0
openai-whisper==20231117
packaging==24.1
pillow==10.3.0
proglog==0.1.10
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sympy==1.12.1
tbb==2021.12.0
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
